% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getNbCases}
\alias{getNbCases}
\title{Get Number of cases of MMRF-COMMPASS project from GDC}
\usage{
getNbCases(data.category)
}
\arguments{
\item{data.category}{A  GDC project data category}
}
\description{
Get Number of cases of MMRF-COMMPASS project from GDC
}
\examples{
\dontrun{
getNbCases("Clinical")
}
}
