% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getNbFiles}
\alias{getNbFiles}
\title{Get Number of files in MMRF-COMMPASS project}
\usage{
getNbFiles(data.category)
}
\arguments{
\item{data.category}{A  GDC project data category}

\item{legacy}{Select between Harmonized or Legacy database}
}
\description{
Get Number of files in MMRF-COMMPASS project
}
\examples{
\dontrun{
getNbFiles("Clinical")
}
}
