% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFanalyze.R
\name{MMRF_FPKM_filterByGenes}
\alias{MMRF_FPKM_filterByGenes}
\title{#' @title MMRF_get_barcodeByGenes.}
\usage{
MMRF_FPKM_filterByGenes(dataGE, genes)
}
\arguments{
\item{dataGE}{is a matrix of Gene expression (genes in rows, samples in cols) from MMRFprepare}

\item{geneList}{a character vector of the genes symbols.}
}
\value{
a subset of the Gene expression matrix
}
\description{
MMRF_get_barcodeByGenes returns a subset of the Gene expression matrix (genes in rows, samples in cols) from MMRFprepare.
}
