% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{MMRF_getProjectSummary}
\alias{MMRF_getProjectSummary}
\title{Get MMRF-COMPASS Project Summary from GDC}
\usage{
MMRF_getProjectSummary()
}
\description{
Get MMRF-COMPASS Project Summary from GDC
}
\examples{
\dontrun{
getProjectSummary()
}
}
