% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFClinical.R
\name{MMRFfilterBarcodesByTypes}
\alias{MMRFfilterBarcodesByTypes}
\title{Retrieve barcode filtered by tissue or sample type.}
\usage{
MMRFfilterBarcodesByTypes(clin.bio, sampletype)
}
\arguments{
\item{clin.bio}{is data frame with the clinical information (A valid type is Biospecimen)}

\item{sampletype}{a character vector indicating sample type}

\item{tissuetype}{a character vector indicating tissue type
Example:
\tabular{ll}{
TB \tab   Primary Blood Derived Cancer-Peripheral Blood \cr
TRBM \tab Recurrent Blood Derived Cancer-Bone Marrow \cr
TBM \tab  Primary Blood Derived Cancer-Bone Marrow \cr
NB \tab   Blood Derived Normal \cr
TRB \tab Recurrent Blood Derived Cancer - Peripheral Blood \cr
}}
}
\value{
a list of barcode filtered by  sample or  tissue type selected
}
\description{
MMRFfilterBarcodesByTypes returns the barcode of samples filtered on theses types.
}
\examples{
clin.bio<-MMRFquery_clinic(type = "biospecimen")
 selection of Normal samples
 MMRFfilterBarcodesByTypes(clin.bio,tissuetype="Normal")
 selection of Tumor samples
 MMRFfilterBarcodesByTypes(clin.bio)
 selection of barcodes with Tumor tissue and sample type "Recurrent Blood Derived Cancer - Bone Marrow" (TRBM)
 MMRFfilterBarcodesByTypes(clin.bio,"TRBM","Tumor")
}
