% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFClinical.R
\name{MMRFquery_clinic}
\alias{MMRFquery_clinic}
\title{Get GDC clinical data}
\usage{
MMRFquery_clinic(type = "clinical", save.csv = FALSE)
}
\arguments{
\item{type}{A valid type. Options "clinical", "Biospecimen"  (see list with getGDCprojects()$project_id)]}

\item{save.csv}{Write clinical information into a csv document}

\item{"MMRF-COMMPASS"}{project
\itemize{
\item{ MMRF-COMMPASS }
}}
}
\value{
A data frame with the clinical information
}
\description{
MMRFquery_clinical will download and prepare all clinical information from the API
as the one with using the button from each project
}
\examples{
clin.mm<-MMRFquery_clinic(type = "clinical", save.csv = TRUE)
clin.mm<-MMRFquery_clinic(type = "biospecimen", save.csv = TRUE)
\dontrun{
clin <- MMRFquery_clinical(type = "clinical")
}
}
